add mouse fix parse

+++ bonus/src_bonus/moveUpDown_bonus.c
@@ -6,7 +6,7 @@
 /*   By: aaitbelh <aaitbelh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/18 11:33:40 by aaitbelh          #+#    #+#             */
-/*   Updated: 2022/07/29 13:41:51 by aaitbelh         ###   ########.fr       */
+/*   Updated: 2022/07/31 09:40:39 by aaitbelh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	move_up(t_game *game)
 {
-	float	tmp_x;
-	float	tmp_y;
+	double	tmp_x;
+	double	tmp_y;
 
 	tmp_x = game->player->x + game->player->dirx * game->player->move_speed;
 	tmp_y = game->player->y + game->player->diry * game->player->move_speed;
@@ -32,8 +32,8 @@ void	move_up(t_game *game)
 
 void	move_down(t_game *game)
 {
-	float	tmp_x;
-	float	tmp_y;
+	double	tmp_x;
+	double	tmp_y;
 
 	tmp_x = game->player->x - game->player->dirx * game->player->move_speed;
 	tmp_y = game->player->y - game->player->diry * game->player->move_speed;
